+++ Backend/src/main/java/com/example/backend/user/service/UserService.java
@@ -1,13 +1,17 @@
 package com.example.backend.user.service;
 
 
+import com.example.backend.companies.model.Company;
+import com.example.backend.project.model.Project;
 import com.example.backend.user.controller.dto.UserDTO;
+import com.example.backend.user.model.AUTHORITY;
 import com.example.backend.user.model.AppUser;
 import com.example.backend.user.repo.UserRepository;
+import com.example.backend.util.Helper;
+import com.example.backend.util.exceptions.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.HashSet;
+import javax.naming.AuthenticationException;
 import java.util.List;
 import java.util.Optional;
 
@@ -29,24 +33,25 @@ public class UserService {
         else return userRepository.findByLoginEmail(email).get(0);
     }
 
-    public AppUser addUser(UserDTO userDTO){
-        AppUser appUser = new AppUser(
-                userDTO.getLoginEmail(),
-                userDTO.getAuthority(),
-                userDTO.getFirstName(),
-                userDTO.getLastName(),
-                userDTO.getNotificationEmail(),
-                userDTO.getDescription(),
-                userDTO.getNickname()
-        );
-        return userRepository.save(appUser);
+    public AppUser addUser(UserDTO userDTO, AppUser user) throws AuthenticationException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        if (!List.of(AUTHORITY.ADMINISTRATOR).contains(user.getAuthority())) throw new AuthenticationException("You do not have permission to execute this command.");
+
+        return userRepository.save(userDTO.toAppUser());
     }
 
-    public void deleteUser(Long id){
+    public void deleteUser(Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        Helper.checkUserAuthorities(user, List.of(AUTHORITY.ADMINISTRATOR));
+
+        AppUser appUser = Helper.getValue(userRepository.findById(id), "User under id " + id + " not found.");
+        Helper.checkSoftLocked(appUser);
+
         userRepository.deleteById(id);
     }
 
-    public List<AppUser> findAll(){
+    public List<AppUser> findAll(AppUser user) throws AuthenticationException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+
         return userRepository.findAll();
     }
 
@@ -55,12 +60,31 @@ public class UserService {
         return i > 0;
     }
 
-    public Optional<AppUser> findById(Long id){
-        return userRepository.findById(id);
+    public AppUser findById(Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        if (user == null) throw new AuthenticationException("You do not have permission to access CDB.");
+        Optional<AppUser> appUserOpt = userRepository.findById(id);
+        if (appUserOpt.isPresent()) return appUserOpt.get();
+        else throw new EntityNotFoundException("User with id " + id + "not found.");
     }
 
-    public AppUser updateUser(UserDTO userDTO, Long id){
-        AppUser appUser = userRepository.findById(id).get();
+    public Boolean softLockUser(AppUser user, Long id) throws AuthenticationException, EntityNotFoundException
+    {
+        Helper.checkUserAuthorities(user, List.of(AUTHORITY.ADMINISTRATOR));
+        Helper.checkSoftLocked(user);
+
+        String errorMessage = "User with id " + id + " does not exist";
+        AppUser appUser = Helper.getValue(userRepository.findById(id), errorMessage);
+        boolean newSoftLocked = appUser.getSoftLocked() == null || !appUser.getSoftLocked();
+        appUser.setSoftLocked(newSoftLocked);
+        userRepository.save(appUser);
+        return newSoftLocked;
+    }
+
+    public AppUser updateUser(UserDTO userDTO, Long id, AppUser user) throws AuthenticationException, EntityNotFoundException {
+        Helper.checkUserAuthorities(user, List.of(AUTHORITY.ADMINISTRATOR));
+
+        AppUser appUser = Helper.getValue(userRepository.findById(id), "User under id " + id + " not found.");
+        Helper.checkSoftLocked(appUser);
 
         appUser.setLoginEmail(userDTO.getLoginEmail());
         appUser.setAuthority(userDTO.getAuthority());
